refactor: hydraulics provider
